refactor: rename `migrate-system-types` to `migrate-ontology-types`
