+++ apps/hash-api/src/graph/migrate-ontology-types/util.ts
